ничего не понял
