+++ index.html
@@ -33,7 +33,8 @@
     </div>
   </div>
   </div>
-  <script src="script.js"></script>
+  <!-- <script src="./js/script.js"></script> -->
+  <script type="module" src="js/script.js"></script>
 </body>
 
 <script>
