+++ styles.css
@@ -180,8 +180,7 @@ body {
     background: radial-gradient(
       75.42% 75.42% at 50% 42.37%,
       rgba(53, 53, 53, 0) 22.92%,
-      #7334ea 100%
-    );
+      #7334ea 100%);
     filter: drop-shadow(0px 20px 67px rgba(0, 0, 0, 0.08));
     border-radius: 20px;
     padding: 20px;
